better auth flow
added some more nextjs pages to test different auth + rendering
methods.

- pages/index.js - statically served page
- pages/login.js - server rendered page. if youre logged in, you get
  routed to the home page
- pages/profile.js - server rendered page. if youre not logged in, you
  get routed to the login page

also added chakra just to see if it was easy to use or not. it is.

there is a global UserContextProvider that wraps the entire app.

on the first render, it will set the session in state, as well as
registers any auth state changes. if the user logs in, then a session
gets set via an API to set cookies. if the user logs out, then a session
gets cleared.